+++ packages/core/src/generators/vue/vue.ts
@@ -1,20 +1,11 @@
 import dedent from 'dedent';
-import json5 from 'json5';
-import { types } from '@babel/core';
 import { format } from 'prettier/standalone';
 import { collectCss } from '../../helpers/styles/collect-css';
 import { fastClone } from '../../helpers/fast-clone';
-import {
-  stringifyContextValue,
-  getStateObjectStringFromComponent,
-} from '../../helpers/get-state-object-string';
 import { mapRefs } from '../../helpers/map-refs';
-import { checkIsComponentImport, renderPreComponent } from '../../helpers/render-imports';
-import { stripStateAndPropsRefs } from '../../helpers/strip-state-and-props-refs';
+import { renderPreComponent } from '../../helpers/render-imports';
 import { getProps } from '../../helpers/get-props';
-import { selfClosingTags } from '../../parsers/jsx';
-import { extendedHook, MitosisComponent } from '../../types/mitosis-component';
-import { ForNode, MitosisNode } from '../../types/mitosis-node';
+import { MitosisComponent } from '../../types/mitosis-component';
 import {
   Plugin,
   runPostCodePlugins,
@@ -22,296 +13,21 @@ import {
   runPreCodePlugins,
   runPreJsonPlugins,
 } from '../../modules/plugins';
-import isChildren from '../../helpers/is-children';
 import { stripMetaProperties } from '../../helpers/strip-meta-properties';
-import { removeSurroundingBlock } from '../../helpers/remove-surrounding-block';
 import { isMitosisNode } from '../../helpers/is-mitosis-node';
 import traverse from 'traverse';
-import { getComponentsUsed } from '../../helpers/get-components-used';
-import { kebabCase, pickBy, size, uniq } from 'lodash';
-import { replaceIdentifiers } from '../../helpers/replace-identifiers';
-import { filterEmptyTextNodes } from '../../helpers/filter-empty-text-nodes';
+import { pickBy, size, uniq } from 'lodash';
 import { processHttpRequests } from '../../helpers/process-http-requests';
 import { TranspilerGenerator } from '../../types/transpiler';
-import { GETTER } from '../../helpers/patterns';
 import { OmitObj } from '../../helpers/typescript';
 import { pipe } from 'fp-ts/lib/function';
-import { getCustomImports } from '../../helpers/get-custom-imports';
-import { isSlotProperty, stripSlotPrefix, replaceSlotsInString } from '../../helpers/slots';
-import { PropsDefinition, DefaultProps } from 'vue/types/options';
+import { isSlotProperty } from '../../helpers/slots';
 import { FUNCTION_HACK_PLUGIN } from '../helpers/functions';
-import { babelTransformExpression } from '../../helpers/babel-transform';
-import { checkIsDefined } from '../../helpers/nullable';
-import {
-  addBindingsToJson,
-  addPropertiesToJson,
-  encodeQuotes,
-  getOnUpdateHookName,
-  invertBooleanExpression,
-  renameMitosisComponentsToKebabCase,
-} from './helpers';
+import { getOnUpdateHookName, renameMitosisComponentsToKebabCase } from './helpers';
 import { ToVueOptions, VueVersionOpt } from './types';
-import { processBinding } from './helpers';
-
-const SPECIAL_PROPERTIES = {
-  V_IF: 'v-if',
-  V_FOR: 'v-for',
-  V_ELSE: 'v-else',
-  V_ELSE_IF: 'v-else-if',
-} as const;
-
-type BlockRenderer = (json: MitosisNode, options: ToVueOptions, scope?: Scope) => string;
-
-const NODE_MAPPERS: {
-  [key: string]: BlockRenderer | undefined;
-} = {
-  Fragment(json, options, scope) {
-    if (options.vueVersion === 2 && scope?.isRootNode) {
-      throw new Error('Vue 2 template should have a single root element');
-    }
-    return json.children.map((item) => blockToVue(item, options)).join('\n');
-  },
-  For(_json, options) {
-    const json = _json as ForNode;
-    const keyValue = json.bindings.key || { code: 'index' };
-    const forValue = `(${json.scope.forName}, index) in ${stripStateAndPropsRefs(
-      json.bindings.each?.code,
-    )}`;
-
-    if (options.vueVersion >= 3) {
-      // TODO: tmk key goes on different element (parent vs child) based on Vue 2 vs Vue 3
-      return `<template :key="${encodeQuotes(keyValue?.code || 'index')}" v-for="${encodeQuotes(
-        forValue,
-      )}">
-        ${json.children.map((item) => blockToVue(item, options)).join('\n')}
-      </template>`;
-    }
-    // Vue 2 can only handle one root element
-    const firstChild = json.children.filter(filterEmptyTextNodes)[0] as MitosisNode | undefined;
-
-    // Edge-case for when the parent is a `Show`, we need to pass down the `v-if` prop.
-    const jsonIf = json.properties[SPECIAL_PROPERTIES.V_IF];
-
-    return firstChild
-      ? pipe(
-          firstChild,
-          addBindingsToJson({ key: keyValue }),
-          addPropertiesToJson({
-            [SPECIAL_PROPERTIES.V_FOR]: forValue,
-            ...(jsonIf ? { [SPECIAL_PROPERTIES.V_IF]: jsonIf } : {}),
-          }),
-          (block) => blockToVue(block, options),
-        )
-      : '';
-  },
-  Show(json, options, scope) {
-    const ifValue = replaceSlotsInString(
-      stripStateAndPropsRefs(json.bindings.when?.code),
-      (slotName) => `$slots.${slotName}`,
-    );
-
-    const defaultShowTemplate = `
-    <template ${SPECIAL_PROPERTIES.V_IF}="${encodeQuotes(ifValue)}">
-      ${json.children.map((item) => blockToVue(item, options)).join('\n')}
-    </template>
-    ${
-      isMitosisNode(json.meta.else)
-        ? `
-        <template ${SPECIAL_PROPERTIES.V_ELSE}>
-          ${blockToVue(json.meta.else, options)}
-        </template>`
-        : ''
-    }
-    `;
-
-    switch (options.vueVersion) {
-      case 3:
-        return defaultShowTemplate;
-      case 2:
-        // if it is not the root node, the default show template can be used
-        // as Vue 2 only has this limitation at root level
-        if (!scope?.isRootNode) {
-          return defaultShowTemplate;
-        }
-
-        const children = json.children.filter(filterEmptyTextNodes);
-        // Vue 2 can only handle one root element, so we just take the first one.
-        // TO-DO: warn user of multi-children Show.
-        const firstChild = children[0] as MitosisNode | undefined;
-        const elseBlock = json.meta.else;
-
-        const hasShowChild = firstChild?.name === 'Show';
-        const childElseBlock = firstChild?.meta.else;
-
-        const allShowChildrenWithoutElse = children.every((x) => x.name === 'Show' && !x.meta.else);
-
-        if (allShowChildrenWithoutElse && isMitosisNode(elseBlock)) {
-          /**
-           * This is when we mimic an if-else chain by only providing `Show` elements as children, none of which have an `else` block
-           *
-           * <show when={foo} else={else-1}>
-           *  <show when={bar}> <bar-code> </show>
-           *  <show when={x}> <x-code> </show>
-           *  <show when={y}> <y-code> </show>
-           * </show>
-           *
-           * What we want in this case is:
-           *
-           * <else-1 if={!foo} />
-           * <bar-code v-else-if={bar} />
-           * <x-code v-else-if={x} />
-           * <y-code v-else />
-           */
-          const ifString = pipe(
-            elseBlock,
-            addPropertiesToJson({ [SPECIAL_PROPERTIES.V_IF]: invertBooleanExpression(ifValue) }),
-            (block) => blockToVue(block, options),
-          );
-
-          const childrenStrings = children.map((child, idx) => {
-            const isLast = idx === children.length - 1;
-
-            const innerBlock = child.children.filter(filterEmptyTextNodes)[0];
-
-            if (!isLast) {
-              const childIfValue = pipe(child.bindings.when?.code, stripStateAndPropsRefs);
-              const elseIfString = pipe(
-                innerBlock,
-                addPropertiesToJson({ [SPECIAL_PROPERTIES.V_ELSE_IF]: childIfValue }),
-                (block) => blockToVue(block, options),
-              );
-
-              return elseIfString;
-            } else {
-              const elseString = pipe(
-                innerBlock,
-                addPropertiesToJson({ [SPECIAL_PROPERTIES.V_ELSE]: '' }),
-                (block) => blockToVue(block, options),
-              );
-
-              return elseString;
-            }
-          });
-
-          return `
-            ${ifString}
-            ${childrenStrings.join('\n')}
-          `;
-        } else if (
-          firstChild &&
-          isMitosisNode(elseBlock) &&
-          hasShowChild &&
-          isMitosisNode(childElseBlock)
-        ) {
-          /**
-           * This is special edge logic to handle 2 nested Show elements in Vue 2.
-           * We need to invert the logic to make it work, due to no-template-root-element limitations in Vue 2.
-           *
-           * <show when={foo} else={else-1}>
-           *  <show when={bar}> <bar-code> </show>
-           *
-           *  <show when={x}> <x-code> </show>
-           *
-           *  <show when={y}> <y-code> </show>
-           * </show>
-           *
-           *
-           *
-           *
-           * foo: true && bar: true => if-code
-           * foo: true && bar: false => else-2
-           * foo: false && bar: true?? => else-1
-           *
-           *
-           * map to:
-           *
-           * <else-1 if={!foo} />
-           * <else-2 v-else-if={!bar} />
-           * <if-code v-else />
-           *
-           */
-          const ifString = pipe(
-            elseBlock,
-            addPropertiesToJson({ [SPECIAL_PROPERTIES.V_IF]: invertBooleanExpression(ifValue) }),
-            (block) => blockToVue(block, options),
-          );
-
-          const childIfValue = pipe(
-            firstChild.bindings.when?.code,
-            stripStateAndPropsRefs,
-            invertBooleanExpression,
-          );
-          const elseIfString = pipe(
-            childElseBlock,
-            addPropertiesToJson({ [SPECIAL_PROPERTIES.V_ELSE_IF]: childIfValue }),
-            (block) => blockToVue(block, options),
-          );
-
-          const firstChildOfFirstChild = firstChild.children.filter(filterEmptyTextNodes)[0] as
-            | MitosisNode
-            | undefined;
-          const elseString = firstChildOfFirstChild
-            ? pipe(
-                firstChildOfFirstChild,
-                addPropertiesToJson({ [SPECIAL_PROPERTIES.V_ELSE]: '' }),
-                (block) => blockToVue(block, options),
-              )
-            : '';
-
-          return `
-
-            ${ifString}
-
-            ${elseIfString}
-
-            ${elseString}
-
-          `;
-        } else {
-          const ifString = firstChild
-            ? pipe(
-                firstChild,
-                addPropertiesToJson({ [SPECIAL_PROPERTIES.V_IF]: ifValue }),
-                (block) => blockToVue(block, options),
-              )
-            : '';
-
-          const elseString = isMitosisNode(elseBlock)
-            ? pipe(elseBlock, addPropertiesToJson({ [SPECIAL_PROPERTIES.V_ELSE]: '' }), (block) =>
-                blockToVue(block, options),
-              )
-            : '';
-
-          return `
-                    ${ifString}
-                    ${elseString}
-                  `;
-        }
-    }
-  },
-  Slot(json, options) {
-    if (!json.bindings.name) {
-      const key = Object.keys(json.bindings).find(Boolean);
-      if (!key) return '<slot />';
-
-      return `
-        <template #${key}>
-        ${json.bindings[key]?.code}
-        </template>
-      `;
-    }
-    const strippedTextCode = stripStateAndPropsRefs(json.bindings.name.code);
-
-    return `<slot name="${stripSlotPrefix(strippedTextCode).toLowerCase()}">${json.children
-      ?.map((item) => blockToVue(item, options))
-      .join('\n')}</slot>`;
-  },
-};
-
-// TODO: Maybe in the future allow defining `string | function` as values
-const BINDING_MAPPERS: { [key: string]: string | undefined } = {
-  innerHTML: 'v-html',
-};
+import { generateOptionsApiScript } from './optionsApi';
+import { generateCompositionApiScript } from './compositionApi';
+import { blockToVue } from './blocks';
 
 // Transform <foo.bar key="value" /> to <component :is="foo.bar" key="value" />
 function processDynamicComponents(json: MitosisComponent, _options: ToVueOptions) {
@@ -337,164 +53,6 @@ function processForKeys(json: MitosisComponent, _options: ToVueOptions) {
       }
     }
   });
-}
-
-const stringifyBinding =
-  (node: MitosisNode) =>
-  ([key, value]: [string, { code: string; arguments?: string[] } | undefined]) => {
-    if (key === '_spread') {
-      return '';
-    } else if (key === 'class') {
-      return ` :class="_classStringToObject(${stripStateAndPropsRefs(value?.code, {
-        replaceWith: '',
-      })})" `;
-      // TODO: support dynamic classes as objects somehow like Vue requires
-      // https://vuejs.org/v2/guide/class-and-style.html
-    } else {
-      // TODO: proper babel transform to replace. Util for this
-      const useValue = stripStateAndPropsRefs(value?.code);
-
-      if (key.startsWith('on')) {
-        const { arguments: cusArgs = ['event'] } = value!;
-        let event = key.replace('on', '').toLowerCase();
-        if (event === 'change' && node.name === 'input') {
-          event = 'input';
-        }
-        const isAssignmentExpression = useValue.includes('=');
-        const valueWRenamedEvent = replaceIdentifiers({
-          code: useValue,
-          from: cusArgs[0],
-          to: '$event',
-        });
-
-        // TODO: proper babel transform to replace. Util for this
-        if (isAssignmentExpression) {
-          return ` @${event}="${encodeQuotes(removeSurroundingBlock(valueWRenamedEvent))}" `;
-        } else {
-          return ` @${event}="${encodeQuotes(
-            removeSurroundingBlock(removeSurroundingBlock(valueWRenamedEvent)),
-          )}" `;
-        }
-      } else if (key === 'ref') {
-        return ` ref="${encodeQuotes(useValue)}" `;
-      } else if (BINDING_MAPPERS[key]) {
-        return ` ${BINDING_MAPPERS[key]}="${encodeQuotes(useValue.replace(/"/g, "\\'"))}" `;
-      } else {
-        return ` :${key}="${encodeQuotes(useValue)}" `;
-      }
-    }
-  };
-
-interface Scope {
-  isRootNode?: boolean;
-}
-
-export const blockToVue: BlockRenderer = (node, options, scope) => {
-  const nodeMapper = NODE_MAPPERS[node.name];
-  if (nodeMapper) {
-    return nodeMapper(node, options, scope);
-  }
-
-  if (isChildren(node)) {
-    return `<slot/>`;
-  }
-
-  if (node.name === 'style') {
-    // Vue doesn't allow <style>...</style> in templates, but does support the synonymous
-    // <component is="'style'">...</component>
-    node.name = 'component';
-    node.bindings.is = { code: "'style'" };
-  }
-
-  if (node.properties._text) {
-    return `${node.properties._text}`;
-  }
-
-  const textCode = node.bindings._text?.code;
-  if (textCode) {
-    const strippedTextCode = stripStateAndPropsRefs(textCode);
-    if (isSlotProperty(strippedTextCode)) {
-      return `<slot name="${stripSlotPrefix(strippedTextCode).toLowerCase()}"/>`;
-    }
-    return `{{${strippedTextCode}}}`;
-  }
-
-  let str = '';
-
-  str += `<${node.name} `;
-
-  if (node.bindings._spread?.code) {
-    str += `v-bind="${encodeQuotes(stripStateAndPropsRefs(node.bindings._spread.code as string))}"`;
-  }
-
-  for (const key in node.properties) {
-    const value = node.properties[key];
-
-    if (key === 'className') {
-      continue;
-    } else if (key === SPECIAL_PROPERTIES.V_ELSE) {
-      str += ` ${key} `;
-    } else if (typeof value === 'string') {
-      str += ` ${key}="${encodeQuotes(value)}" `;
-    }
-  }
-
-  const stringifiedBindings = Object.entries(node.bindings)
-    .map(([k, v]) =>
-      stringifyBinding(node)([k, v] as [
-        string,
-        { code: string; arguments?: string[] } | undefined,
-      ]),
-    )
-    .join('');
-
-  str += stringifiedBindings;
-
-  if (selfClosingTags.has(node.name)) {
-    return str + ' />';
-  }
-
-  str += '>';
-  if (node.children) {
-    str += node.children.map((item) => blockToVue(item, options)).join('');
-  }
-
-  return str + `</${node.name}>`;
-};
-
-function getContextInjectString(component: MitosisComponent, options: ToVueOptions) {
-  let str = '{';
-
-  for (const key in component.context.get) {
-    str += `
-      ${key}: "${encodeQuotes(component.context.get[key].name)}",
-    `;
-  }
-
-  str += '}';
-  return str;
-}
-
-function getContextProvideString(component: MitosisComponent, options: ToVueOptions) {
-  let str = '{';
-
-  for (const key in component.context.set) {
-    const { ref, value, name } = component.context.set[key];
-    str += `
-      ${name}: ${
-      value
-        ? stringifyContextValue(value, {
-            valueMapper: (code) => stripStateAndPropsRefs(code, { replaceWith: '_this.' }),
-          })
-        : ref
-        ? stripStateAndPropsRefs(ref, { replaceWith: '_this.' })
-        : null
-    },
-    `;
-  }
-
-  str += '}';
-  return str;
 }
 
 /**
@@ -547,409 +105,6 @@ const mergeOptions = (
   ...b,
   plugins: [...pluginsA, ...pluginsB],
 });
-
-const generateComponentImport =
-  (options: ToVueOptions) =>
-  (componentName: string): string => {
-    if (options.vueVersion >= 3 && options.asyncComponentImports) {
-      return `'${componentName}': defineAsyncComponent(${componentName})`;
-    } else {
-      return `'${componentName}': ${componentName}`;
-    }
-  };
-
-const generateComponents = (componentsUsed: string[], options: ToVueOptions): string => {
-  if (componentsUsed.length === 0) {
-    return '';
-  } else {
-    return `components: { ${componentsUsed.map(generateComponentImport(options)).join(',')} },`;
-  }
-};
-
-const appendToDataString = ({
-  dataString,
-  newContent,
-}: {
-  dataString: string;
-  newContent: string;
-}) => dataString.replace(/}$/, `${newContent}}`);
-
-function generateOptionsApiScript(
-  component: MitosisComponent,
-  options: ToVueOptions,
-  path: string | undefined,
-  template: string,
-  props: string[],
-  onUpdateWithDeps: extendedHook[],
-  onUpdateWithoutDeps: extendedHook[],
-) {
-  const { exports: localExports } = component;
-  const localVarAsData: string[] = [];
-  const localVarAsFunc: string[] = [];
-  if (localExports) {
-    Object.keys(localExports).forEach((key) => {
-      if (localExports[key].usedInLocal) {
-        if (localExports[key].isFunction) {
-          localVarAsFunc.push(key);
-        } else {
-          localVarAsData.push(key);
-        }
-      }
-    });
-  }
-
-  let dataString = getStateObjectStringFromComponent(component, {
-    data: true,
-    functions: false,
-    getters: false,
-  });
-
-  // Append refs to data as { foo, bar, etc }
-  dataString = appendToDataString({
-    dataString,
-    newContent: getCustomImports(component).join(','),
-  });
-
-  if (localVarAsData.length) {
-    dataString = appendToDataString({ dataString, newContent: localVarAsData.join(',') });
-  }
-
-  const getterString = getStateObjectStringFromComponent(component, {
-    data: false,
-    getters: true,
-    functions: false,
-    valueMapper: (code) =>
-      processBinding({ code: code.replace(GETTER, ''), options, json: component }),
-  });
-
-  let functionsString = getStateObjectStringFromComponent(component, {
-    data: false,
-    getters: false,
-    functions: true,
-    valueMapper: (code) => processBinding({ code, options, json: component }),
-  });
-
-  const includeClassMapHelper = template.includes('_classStringToObject');
-
-  if (includeClassMapHelper) {
-    functionsString = functionsString.replace(
-      /}\s*$/,
-      `_classStringToObject(str) {
-        const obj = {};
-        if (typeof str !== 'string') { return obj }
-        const classNames = str.trim().split(/\\s+/);
-        for (const name of classNames) {
-          obj[name] = true;
-        }
-        return obj;
-      }  }`,
-    );
-  }
-
-  if (localVarAsFunc.length) {
-    functionsString = functionsString.replace(/}\s*$/, `${localVarAsFunc.join(',')}}`);
-  }
-
-  // Component references to include in `component: { YourComponent, ... }
-  const componentsUsedInTemplate = Array.from(getComponentsUsed(component))
-    .filter((name) => name.length && !name.includes('.') && name[0].toUpperCase() === name[0])
-    // Strip out components that compile away
-    .filter((name) => !['For', 'Show', 'Fragment', 'Slot', component.name].includes(name));
-
-  // get default imports from component files
-  const importedComponents = component.imports
-    .filter(checkIsComponentImport)
-    .map((imp) => Object.entries(imp.imports).find(([_, value]) => value === 'default')?.[0])
-    .filter(checkIsDefined);
-
-  const componentsUsed = uniq([...componentsUsedInTemplate, ...importedComponents]);
-
-  let propsDefinition: PropsDefinition<DefaultProps> = Array.from(props).filter(
-    (prop) => prop !== 'children' && prop !== 'class',
-  );
-
-  // add default props (if set)
-  if (component.defaultProps) {
-    propsDefinition = propsDefinition.reduce(
-      (propsDefinition: DefaultProps, curr: string) => (
-        (propsDefinition[curr] = component.defaultProps?.hasOwnProperty(curr)
-          ? { default: component.defaultProps[curr] }
-          : {}),
-        propsDefinition
-      ),
-      {},
-    );
-  }
-
-  return `
-        export default {
-        ${
-          !component.name
-            ? ''
-            : `name: '${
-                path && options.namePrefix?.(path) ? options.namePrefix?.(path) + '-' : ''
-              }${kebabCase(component.name)}',`
-        }
-        ${generateComponents(componentsUsed, options)}
-        ${props.length ? `props: ${json5.stringify(propsDefinition)},` : ''}
-        ${
-          dataString.length < 4
-            ? ''
-            : `
-        data: () => (${dataString}),
-        `
-        }
-
-        ${
-          size(component.context.set)
-            ? `provide() {
-                const _this = this;
-                return ${getContextProvideString(component, options)}
-              },`
-            : ''
-        }
-        ${
-          size(component.context.get)
-            ? `inject: ${getContextInjectString(component, options)},`
-            : ''
-        }
-        ${
-          component.hooks.onInit?.code
-            ? `created() {
-                ${processBinding({ code: component.hooks.onInit.code, options, json: component })}
-              },`
-            : ''
-        }
-        ${
-          component.hooks.onMount?.code
-            ? `mounted() {
-                ${processBinding({ code: component.hooks.onMount.code, options, json: component })}
-              },`
-            : ''
-        }
-        ${
-          onUpdateWithoutDeps.length
-            ? `updated() {
-            ${onUpdateWithoutDeps
-              .map((hook) => processBinding({ code: hook.code, options, json: component }))
-              .join('\n')}
-          },`
-            : ''
-        }
-        ${
-          onUpdateWithDeps.length
-            ? `watch: {
-            ${onUpdateWithDeps
-              .map(
-                (hook, index) =>
-                  `${getOnUpdateHookName(index)}() {
-                  ${processBinding({ code: hook.code, options, json: component })}
-                  }
-                `,
-              )
-              .join(',')}
-          },`
-            : ''
-        }
-        ${
-          component.hooks.onUnMount
-            ? `unmounted() {
-                ${processBinding({
-                  code: component.hooks.onUnMount.code,
-                  options,
-                  json: component,
-                })}
-              },`
-            : ''
-        }
-
-        ${
-          getterString.length < 4
-            ? ''
-            : ` 
-          computed: ${getterString},
-        `
-        }
-        ${
-          functionsString.length < 4
-            ? ''
-            : `
-          methods: ${functionsString},
-        `
-        }
-        ${Object.entries(component.meta.vueConfig || {})
-          .map(([k, v]) => `${k}: ${v}`)
-          .join(',')}
-      }`;
-}
-
-const getCompositionPropDefinition = ({
-  options,
-  component,
-  props,
-}: {
-  options: ToVueOptions;
-  component: MitosisComponent;
-  props: string[];
-}) => {
-  let str = 'const props = ';
-
-  if (component.defaultProps) {
-    const generic = options.typescript ? `<${component.propsTypeRef}>` : '';
-    str += `withDefaults(defineProps${generic}(), ${json5.stringify(component.defaultProps)})`;
-  } else if (options.typescript && component.propsTypeRef && component.propsTypeRef !== 'any') {
-    str += `defineProps<${component.propsTypeRef}>()`;
-  } else {
-    str += `defineProps(${json5.stringify(props)})`;
-  }
-  return str;
-};
-
-function appendValueToRefs(input: string, component: MitosisComponent, options: ToVueOptions) {
-  const refKeys = Object.keys(pickBy(component.state, (i) => i?.type === 'property'));
-
-  let output = processBinding({ code: input, options, json: component, includeProps: false });
-
-  return babelTransformExpression(output, {
-    Identifier(path: babel.NodePath<babel.types.Identifier>) {
-      if (
-        !(types.isFunctionDeclaration(path.parent) && path.parent.id === path.node) &&
-        !types.isCallExpression(path.parent) &&
-        (!types.isMemberExpression(path.parent) || types.isThisExpression(path.parent.object)) &&
-        path.parentPath.listKey !== 'arguments' &&
-        path.parentPath.listKey !== 'params' &&
-        refKeys.includes(path.node.name)
-      ) {
-        path.replaceWith(types.identifier(`${path.node.name}.value`));
-      }
-    },
-  });
-}
-
-function generateCompositionApiScript(
-  component: MitosisComponent,
-  options: ToVueOptions,
-  template: string,
-  props: Array<string>,
-  onUpdateWithDeps: extendedHook[],
-  onUpdateWithoutDeps: extendedHook[],
-) {
-  let refs = getStateObjectStringFromComponent(component, {
-    data: true,
-    functions: false,
-    getters: false,
-    format: 'variables',
-    valueMapper: (code) => {
-      return processBinding({ code: `ref(${code})`, options, json: component });
-    },
-    keyPrefix: 'const',
-  });
-
-  let methods = getStateObjectStringFromComponent(component, {
-    data: false,
-    getters: false,
-    functions: true,
-    valueMapper: (code) => processBinding({ code, options, json: component, includeProps: false }),
-    format: 'variables',
-  });
-
-  if (template.includes('_classStringToObject')) {
-    methods += ` function _classStringToObject(str) {
-    const obj = {};
-    if (typeof str !== 'string') { return obj }
-    const classNames = str.trim().split(/\\s+/);
-    for (const name of classNames) {
-      obj[name] = true;
-    }
-    return obj;
-    } `;
-  }
-
-  const getterKeys = Object.keys(pickBy(component.state, (i) => i?.type === 'getter'));
-
-  let str = dedent`
-    ${props.length ? getCompositionPropDefinition({ component, props, options }) : ''}
-    ${refs}
-
-    ${Object.keys(component.context.get)
-      ?.map((key) => `const ${key} = inject(${component.context.get[key].name})`)
-      .join('\n')}
-
-    ${Object.keys(component.context.set)
-      ?.map(
-        (key) => `provide(${component.context.set[key].name}, ${component.context.set[key].ref})`,
-      )
-      .join('\n')}
-
-    ${Object.keys(component.refs)
-      ?.map((key) => {
-        if (options.typescript) {
-          return `const ${key} = ref<${component.refs[key].typeParameter}>()`;
-        } else {
-          return `const ${key} = ref(null)`;
-        }
-      })
-      .join('\n')}
-    ${appendValueToRefs(component.hooks.onInit?.code ?? '', component, options)}
-    ${
-      !component.hooks.onMount?.code
-        ? ''
-        : `onMounted(() => { ${appendValueToRefs(
-            component.hooks.onMount.code,
-            component,
-            options,
-          )}})`
-    }
-    ${
-      !component.hooks.onUnMount?.code
-        ? ''
-        : `onMounted(() => { ${appendValueToRefs(
-            component.hooks.onUnMount.code,
-            component,
-            options,
-          )}})`
-    }
-    ${
-      !getterKeys
-        ? ''
-        : getterKeys
-            .map((key) => {
-              const code = component.state[key]?.code?.toString();
-              return !code
-                ? ''
-                : `const ${key} = computed(${appendValueToRefs(
-                    code.replace(key, '').replace('get ()', '() =>'),
-                    component,
-                    options,
-                  )})`;
-            })
-            .join('\n')
-    }
-    ${
-      !onUpdateWithoutDeps?.length
-        ? ''
-        : onUpdateWithoutDeps.map((hook) => {
-            return `onUpdated(() => ${appendValueToRefs(hook.code, component, options)})`;
-          })
-    }
-
-    ${
-      !onUpdateWithDeps?.length
-        ? ''
-        : onUpdateWithDeps.map((hook) => {
-            return appendValueToRefs(
-              `watch(${hook.deps}, (${stripStateAndPropsRefs(hook.deps)}) => { ${hook.code}})\n`,
-              component,
-              options,
-            );
-          })
-    }
-    ${methods?.length ? appendValueToRefs(methods, component, options) : ''}
-  `;
-
-  str = str.replace(/this\./g, ''); // strip this elsewhere (e.g. functions)
-  return str;
-}
 
 const componentToVue: TranspilerGenerator<ToVueOptions> =
   (userOptions = BASE_OPTIONS) =>
